utils_python/requirements.txt
Flask>=2.2
torch>=2.0
Pillow>=9.0
numpy>=1.24
requests>=2.0
# If you have a local package for your VitonHD model, include it too:
# viton-model @ file://../path-to-your-viton-model
